+++ plugin-api/src/main/java/org/hyperledger/besu/plugin/services/TraceService.java
@@ -17,6 +17,7 @@ package org.hyperledger.besu.plugin.services;
 import org.hyperledger.besu.datatypes.Hash;
 import org.hyperledger.besu.evm.worldstate.WorldUpdater;
 import org.hyperledger.besu.plugin.Unstable;
+import org.hyperledger.besu.plugin.data.BlockTraceResult;
 import org.hyperledger.besu.plugin.services.tracer.BlockAwareOperationTracer;
 
 import java.util.function.Consumer;
@@ -29,16 +30,18 @@ public interface TraceService extends BesuService {
    *
    * @param blockNumber the block number
    * @param tracer the tracer (OperationTracer)
+   * @return BlockTraceResult the result of the trace
    */
-  void traceBlock(long blockNumber, BlockAwareOperationTracer tracer);
+  BlockTraceResult traceBlock(long blockNumber, BlockAwareOperationTracer tracer);
 
   /**
    * Traces a block by hash
    *
    * @param hash the block hash
    * @param tracer the tracer (OperationTracer)
+   * @return BlockTraceResult the result of the trace
    */
-  void traceBlock(Hash hash, BlockAwareOperationTracer tracer);
+  BlockTraceResult traceBlock(Hash hash, BlockAwareOperationTracer tracer);
 
   /**
    * Traces range of blocks